remove mixins
