--- starport/ui/src/mixins/initCosmos.js
@@ -1,5 +1,0 @@
-import backendConnection from './backend/initConnection'
-
-export default {
-  mixins: [backendConnection]
-}
